requirements.txt
ipywidgets==7.5.1
pandas==1.1.4
streamlit==0.71.0
Pillow==8.0.1
